fixed first bug

+++ lib/ship_builder.rb
@@ -15,6 +15,7 @@ class ShipBuilder
 
   def create name
     raise 'collection is empty' if empty?
+    raise 'invalid ship name' if !collection.keys.include?(name)
     ship = Ship.new(get_size(name))
     collection.delete(name)
     ship
